Split out bootstrap to not have analyzer dependency.
