+++ build_runner/lib/src/build_plan/build_packages.dart
@@ -2,7 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:build/build.dart';
+import 'package:build/src/asset_id.dart';
+import 'package:build/src/exceptions.dart';
 import 'package:built_collection/built_collection.dart';
 import 'package:graphs/graphs.dart';
 import 'package:meta/meta.dart';
