+++ build_runner/lib/src/io/asset_path_provider.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:build/build.dart';
+import 'package:build/src/asset_id.dart';
 
 import '../constants.dart';
