+++ build_runner/lib/src/logging/build_log_messages.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:build/build.dart' show AssetId;
+import 'package:build/src/asset_id.dart';
 import 'package:built_value/built_value.dart';
 import 'package:logging/logging.dart';
